Make problem local variable

+++ src/lib/platform.c
@@ -12,7 +12,7 @@
 
 void delay(unsigned long milliseconds){
 #ifndef attiny4
-    int i,j = 0;
+    long i,j = 0;
     for(i; i < milliseconds; i++){
         for(j; j < CYCLES_PER_MS; j++){
             asm volatile ("nop");
